refactor(core): extract proc.rs from main.rs
Code move + minimal visibility lifts: 7 route handlers, 1 auth gate
helper, and 3 namespace-allow constants relocate from main.rs into a
new core/src/proc.rs.

Moved (handlers, all pub(crate)):
  root_hint                     /              identity hint
  proc_version                  /proc/version  core version string
  proc_worlds                   /proc/worlds   world listing
  proc_du                       /proc/du       per-world body bytes
  proc_df                       /proc/df       storage / memory / quota
  proc_audit_verify             /proc/audit/{world}/verify
  proc_reserved                 /proc, /proc/* fallback

Moved (helper, module-private):
  require_read                  shared auth gate for proc_du / proc_df

Moved (constants, pub(crate) so inline tests in main.rs can match
the exact ALLOW header values):
  ROOT_ALLOW, PROC_ALLOW, AUDIT_VERIFY_ALLOW

Visibility lifts in main.rs (required for the move to compile):
  VERSION       fn -> pub(crate)   used by proc_version / root_hint
  can_read      fn -> pub(crate)   used by require_read

Module wiring: main.rs adds `mod proc;` and one line to the
crate-root re-export block (`pub(crate) use crate::proc::*;`),
the same pattern PR #98 (response.rs) and PR #99 (path.rs) set.
The route table (lines 483-491 of main.rs) and the inline tests
that reach handlers via super::* keep working without import
churn.

This is NOT a strict pure-mv PR because of the 5 visibility lifts.
All 5 are byte-additions of `pub(crate)` keywords on existing
declarations, with no body change — reviewer can verify in seconds.

Stays in main.rs:
- Core struct, route table, main()
- BlockingSqliteError + audit_append_blocking + world_exists_blocking
  (spawn_blocking scaffold, not introspection)
- handle_get / handle_head / handle_put / handle_post / handle_delete
  (durable-write paths, not /proc/*)

This is PR 3 of the v7 FSM-pipeline cascade. Base = master (PR 0/1/2
already merged via the previous cascade flush).

Stats:
  322 insertions
  260 deletions
  582 total churn
  Insertion-only cognitive surface = 322 (under the 500 budget for
  AI co-author working memory; deletions are byte-equivalent to
  matching insertions, verifiable by side-by-side diff)

cargo fmt --manifest-path core/Cargo.toml -- --check: clean
cargo clippy --manifest-path core/Cargo.toml --all-targets -- -D warnings: clean
cargo test --manifest-path core/Cargo.toml: 100 passed, 0 failed
